# Topic/more to base64 perf