--- a/benchmark/src/main/scala/ScodecBitsBenchmark.scala
+++ b/benchmark/src/main/scala/ScodecBitsBenchmark.scala
@@ -25,6 +25,7 @@ class ScodecBitsBenchmark {
   val byteStringChunks_M = (0L until M).map(b => ByteString(b.toByte)).toList
   val bytes_M = Array.tabulate(M.toInt)(i => i.toByte)
   val bitVector_M = bitChunks_M.foldLeft(BitVector.empty)(_ ++ _)
+  val bitVector_M_compact = bitVector_M.copy
   val byteVector_M = byteChunks_M.foldLeft(ByteVector.empty)(_ ++ _)
   val byteString_M = byteStringChunks_M.foldLeft(ByteString())(_ ++ _)
 
@@ -123,6 +124,10 @@ class ScodecBitsBenchmark {
 
   @Benchmark def toBase64(): String =
     bitVector_M.toBase64
+  @Benchmark def toBase64_compact(): String =
+    bitVector_M_compact.toBase64
   @Benchmark def toBase64_JRE(): String =
     java.util.Base64.getEncoder.encodeToString(bitVector_M.toByteArray)
+  @Benchmark def toBase64_JRE_compact(): String =
+    java.util.Base64.getEncoder.encodeToString(bitVector_M_compact.toByteArray)
 }