`run -i 10 -wi 10 -f1 -t1 .*toBase64.*`

### Before

```
[info] Benchmark                                 Mode  Cnt   Score   Error  Units
[info] ScodecBitsBenchmark.toBase64              avgt   10  20.976 ± 0.790  us/op
[info] ScodecBitsBenchmark.toBase64_JRE          avgt   10   9.855 ± 0.311  us/op
[info] ScodecBitsBenchmark.toBase64_JRE_compact  avgt   10   2.923 ± 0.130  us/op
[info] ScodecBitsBenchmark.toBase64_compact      avgt   10  10.217 ± 0.341  us/op
```

### After

```
[info] Benchmark                                 Mode  Cnt   Score   Error  Units
[info] ScodecBitsBenchmark.toBase64              avgt   10  10.658 ± 0.330  us/op
[info] ScodecBitsBenchmark.toBase64_JRE          avgt   10   9.636 ± 0.338  us/op
[info] ScodecBitsBenchmark.toBase64_JRE_compact  avgt   10   2.935 ± 0.110  us/op
[info] ScodecBitsBenchmark.toBase64_compact      avgt   10   3.559 ± 0.103  us/op
```