@mpilquist 2015-07-14
For the record, I tried inlining the chars array (avoiding the method call to `alphabet.toChar(idx)` for each char) but it had no measurable effect.